motor and encoder working 🏍🏍

+++ node2/src/drivers/adc.h
@@ -5,6 +5,6 @@
 #include <avr/io.h>
 
 void ADC_init();
-uint8_t ADC_read();
+uint16_t ADC_read();
 
 #endif /* ADC_H_ */
